adding database tests for rating repo

+++ src/main/kotlin/com/duda/backbaseassignment/application/rest/model/request/MovieRatingRequest.kt
@@ -1,6 +1,6 @@
 package com.duda.backbaseassignment.application.rest.model.request
 
 data class MovieRatingRequest(
-    val rating: Int,
+    val rating: Short,
     val movieTitle: String
 )
